move script files
moving them to paraphrase directory where they are used
